### 2022-10-27 #110 @whedon
[PRE REVIEW]: RangeEnclosures.jl: A framework to bound function ranges
**Submitting author:** @lucaferranti (<a href="http://orcid.org/0000-0001-5588-0920">Luca Ferranti</a>)
**Repository:** <a href="https://github.com/JuliaReach/RangeEnclosures.jl" target ="_blank">https://github.com/JuliaReach/RangeEnclosures.jl</a>
**Version:**
**Managing EiC:** Mathieu Besançon
