+++ paper/ref.bib
@@ -1,11 +1,11 @@
-@Conference{legat2017sos,
+@inproceedings{legat2017sos,
   author    = {Legat, Beno{\^\i}t and Coey, Chris and Deits, Robin and Huchette, Joey and Perry, Amelia},
   title     = {{Sum-of-squares optimization in Julia}},
   booktitle = {The First Annual JuMP-dev Workshop},
   year      = {2017},
 }
 
-@incollection{benet2019taylormodels,
+@inproceedings{benet2019taylormodels,
   author    = {Benet, Luis and
             Forets, Marcelo and
             Sanders, David P. and
@@ -34,7 +34,9 @@
   journal={Annual Review of Control, Robotics, and Autonomous Systems},
   volume={4},
   number={1},
-  year={2021}
+  year={2021},
+  url={https://doi.org/10.1146/annurev-control-071420-081941},
+  doi={10.1146/annurev-control-071420-081941}
 }
 
 @article{BezansonEKS17,
