### 2022-11-15 #110 @lucaferranti
> Refer to 'paper.log' for details

where can I find the log?
