### 2022-12-12 #110 @odow
@schillic try removing the `xspace` package. It might not be in the docker image that we use to build the LaTeX.

Do you have any suggestions for potential reviewers?
