### 2022-12-12 #110 @odow
Ah. @lucaferranti  you're missing the https://github.com/JuliaCon/JuliaConSubmission.jl/blob/master/paper/.latexmkrc file from the template repository.
