### 2022-12-12 #110 @odow
You've called it `index.latexmkrc`, not `.latexmkrc`:

https://github.com/JuliaReach/RangeEnclosures.jl/blob/schillic/paper/paper/index.latexmkrc
