### 2022-12-12 #110 @odow
@blegat and @dpsanders are you interested in reviewing this paper for the JuliaCon proceedings?

Given the lead-up to Christmas, our deadline for the first round isn't until the end of January, so there's no expectation that you'll start the review now if you are busy. I just need an indication that you're willing, otherwise I will need to find other reviewers. If you're willing, I'll ping you again in early January with a reminder.
