### 2022-12-19 #110 @schillic
(This was a minor change - see https://github.com/JuliaReach/RangeEnclosures.jl/pull/114)
