### 2023-01-10 #110 @lucaferranti
@carstenbauer I am not sure if JSCON whedon has some tweaks compared to the original one. At least in JOSS the bot has had some updates (including renaming to `@editorial-bot`), if you dont have any extra features maybe updating with upstream could help the workflow?
