### 2023-01-23 #122 @whedon
[REVIEW]: RangeEnclosures.jl: A framework to bound function ranges
**Submitting author:** @lucaferranti (<a href="http://orcid.org/0000-0001-5588-0920">Luca Ferranti</a>)
**Repository:** <a href="https://github.com/JuliaReach/RangeEnclosures.jl" target ="_blank">https://github.com/JuliaReach/RangeEnclosures.jl</a>
**Version:** v0.2.4

- [x] **Page limit:** Is the page limit for extended abstracts respected by the submitted document?

- [x] **Page limit:** Is the page limit for extended abstracts respected by the submitted document?
