### 2023-02-06 #122 @odow
I don't think so on the naming front, so long as there is a part in the paper (e.g., the introduction) that still meets the description.
