### 2023-05-02 575040c
update scripts/README
