+++ README.md
@@ -1,10 +1,19 @@
 # RangeEnclosures.jl
 
-[![Build Status](https://github.com/JuliaReach/RangeEnclosures.jl/actions/workflows/ci.yml/badge.svg?branch=master)](https://github.com/JuliaReach/RangeEnclosures.jl/actions/workflows/ci.yml?query=branch%3Amaster)
-[![Docs dev](https://img.shields.io/badge/docs-dev-blue.svg)](http://juliareach.github.io/RangeEnclosures.jl/dev/)
-[![license](https://img.shields.io/github/license/mashape/apistatus.svg?maxAge=2592000)](LICENSE)
-[![Code coverage](http://codecov.io/github/JuliaReach/RangeEnclosures.jl/coverage.svg?branch=master)](https://codecov.io/github/JuliaReach/RangeEnclosures.jl?branch=master)
-[![Join the chat at https://gitter.im/JuliaReach/Lobby](https://badges.gitter.im/JuliaReach/Lobby.svg)](https://gitter.im/JuliaReach/Lobby?utm_source=badge&utm_medium=badge&utm_campaign=pr-badge&utm_content=badge)
+| **Documentation** | **Status** | **Community** | **License** |
+|:-----------------:|:----------:|:-------------:|:-----------:|
+| [![docs-dev][dev-img]][dev-url] | [![CI][ci-img]][ci-url] [![codecov][cov-img]][cov-url] | [![gitter][chat-img]][chat-url] | [![license][lic-img]][lic-url] |
+
+[dev-img]: https://img.shields.io/badge/docs-latest-blue.svg
+[dev-url]: https://juliareach.github.io/RangeEnclosures.jl/dev/
+[ci-img]: https://github.com/JuliaReach/RangeEnclosures.jl/workflows/CI/badge.svg
+[ci-url]: https://github.com/JuliaReach/RangeEnclosures.jl/actions/workflows/ci.yml
+[cov-img]: https://codecov.io/github/JuliaReach/RangeEnclosures.jl/coverage.svg
+[cov-url]: https://app.codecov.io/github/JuliaReach/RangeEnclosures.jl
+[chat-img]: https://img.shields.io/badge/zulip-join_chat-brightgreen.svg
+[chat-url]: https://julialang.zulipchat.com/#narrow/stream/278609-juliareach
+[lic-img]: https://img.shields.io/github/license/mashape/apistatus.svg
+[lic-url]: https://github.com/JuliaReach/RangeEnclosures.jl/blob/master/LICENSE
 
 A [Julia](http://julialang.org) package to bound the range of real-valued functions.
 
