+++ docs/src/about.md
@@ -31,21 +31,12 @@ name of the issue for easier lookup, e.g., `mforets/7`.
 
 ### Unit testing and continuous integration (CI)
 
-This project is synchronized with Travis CI such that each PR gets tested before
-merging (and the build is automatically triggered after each new commit).
-For the maintainability of this project, it is important to understand and fix
-the failing doctests if they exist.
-
-When you modify code in this package, you should make sure that all unit tests
+This project is synchronized with GitHub Actions such that each PR gets tested
+before merging (and the build is automatically triggered after each new commit).
+For the maintainability of this project, it is important to make all unit tests
 pass.
-To run the unit tests locally, you should do:
 
-```
-$ julia --color=yes test/runtests.jl
-```
-
-Alternatively, you can achieve the same from inside the REPL using the following
-command:
+To run the unit tests locally, you can do:
 
 ```julia
 julia> using Pkg
@@ -67,7 +58,7 @@ followed by the name of the type or function.
 For example, the following command will print the documentation of the `enclose`
 function:
 
-```julia
+```
 julia> ?enclose
 ```
 
@@ -83,7 +74,7 @@ or our sources for examples).
 To generate the documentation locally, run `make.jl`, e.g., by executing the
 following command in the terminal:
 
-```
+```bash
 $ julia --color=yes docs/make.jl
 ```
 
