+++ test/runtests.jl
@@ -1,4 +1,5 @@
-using IntervalOptimisation, TaylorModels, Test, RangeEnclosures, SumOfSquares, SDPA, AffineArithmetic
+using IntervalOptimisation, TaylorModels, Test, RangeEnclosures, SumOfSquares, SDPA,
+      AffineArithmetic
 
 using DynamicPolynomials: @polyvar
 
@@ -7,12 +8,13 @@ available_solvers = (NaturalEnclosure(),
                      AffineArithmeticEnclosure(),
                      MooreSkelboeEnclosure(),
                      TaylorModelsEnclosure(),
-                     BranchAndBoundEnclosure(),
-                    )
+                     BranchAndBoundEnclosure())
 
 include("univariate.jl")
 include("multivariate.jl")
 
 using Documenter
 include("../docs/init.jl")
-@time @testset "doctests" begin doctest(RangeEnclosures) end
+@time @testset "doctests" begin
+    doctest(RangeEnclosures)
+end
