+++ src/sdp.jl
@@ -1,6 +1,7 @@
 using .SumOfSquares
 
-function _enclose_sos(sose::SumOfSquaresEnclosure, p::AbstractPolynomialLike, dom::Interval_or_IntervalBox;
+function _enclose_sos(sose::SumOfSquaresEnclosure, p::AbstractPolynomialLike,
+                      dom::Interval_or_IntervalBox;
                       kwargs...)
     x = variables(p)
 
@@ -11,7 +12,7 @@ function _enclose_sos(sose::SumOfSquaresEnclosure, p::AbstractPolynomialLike, do
     # ============
     model = SOSModel(sose.backend, kwargs...)
     @variable(model, γ) # JuMP decision variable
-    @constraint(model, p <= γ, domain=B, maxdegree=sose.order)
+    @constraint(model, p <= γ, domain = B, maxdegree = sose.order)
     @objective(model, Min, γ)
     optimize!(model)
     upper_bound = objective_value(model)
@@ -21,7 +22,7 @@ function _enclose_sos(sose::SumOfSquaresEnclosure, p::AbstractPolynomialLike, do
     # ============
     model = SOSModel(sose.backend, kwargs...)
     @variable(model, γ) # JuMP decision variable
-    @constraint(model, p >= γ, domain=B, maxdegree=sose.order)
+    @constraint(model, p >= γ, domain = B, maxdegree = sose.order)
     @objective(model, Max, γ)
     optimize!(model)
     lower_bound = objective_value(model)
