+++ test/Project.toml
@@ -15,4 +15,4 @@ DynamicPolynomials = "0.3, 0.4"
 IntervalOptimisation = "0.4.1"
 SDPA = "0.2, 0.3, 0.4"
 SumOfSquares = "0.3.6, 0.4, 0.5, 0.6"
-TaylorModels = "0.3, 0.4, 0.5"
+TaylorModels = "0.3, 0.4, 0.5, 0.6"
