### 2023-05-06 2cef5c3
CompatHelper: bump compat for SDPA to 0.5 for package test, (keep existing compat)
