+++ test/Project.toml
@@ -13,6 +13,6 @@ AffineArithmetic = "0.2"
 Documenter = "0.27"
 DynamicPolynomials = "0.3, 0.4"
 IntervalOptimisation = "0.4.1"
-SDPA = "0.2, 0.3, 0.4"
+SDPA = "0.2, 0.3, 0.4, 0.5"
 SumOfSquares = "0.3.6, 0.4, 0.5, 0.6"
 TaylorModels = "0.3, 0.4, 0.5"
