### 2023-06-14 b2cccbb
format docs/make.jl and benchmarks
