+++ docs/make.jl
@@ -2,25 +2,17 @@ using Documenter, RangeEnclosures
 
 include("init.jl")
 
-makedocs(
-    sitename = "RangeEnclosures.jl",
-    modules = [RangeEnclosures],
-    format = Documenter.HTML(
-        prettyurls = get(ENV, "CI", nothing) == "true",
-        assets = ["assets/aligned.css"]),
-    doctest = false,
-    strict = true,
-    pages = [
-        "Home" => "index.md",
-        "Tutorial" => "tutorial.md",
-        "Library" => Any[
-            "Types" => "lib/types.md",
-            "Methods" => "lib/methods.md"],
-        "About" => "about.md"
-    ]
-)
+makedocs(; sitename="RangeEnclosures.jl",
+         modules=[RangeEnclosures],
+         format=Documenter.HTML(; prettyurls=get(ENV, "CI", nothing) == "true",
+                                assets=["assets/aligned.css"]),
+         doctest=false,
+         strict=true,
+         pages=["Home" => "index.md",
+                "Tutorial" => "tutorial.md",
+                "Library" => Any["Types" => "lib/types.md",
+                                 "Methods" => "lib/methods.md"],
+                "About" => "about.md"])
 
-deploydocs(
-    repo = "github.com/JuliaReach/RangeEnclosures.jl.git",
-    push_preview=true
-)
+deploydocs(; repo="github.com/JuliaReach/RangeEnclosures.jl.git",
+           push_preview=true)
