### 2023-06-23 c7f6123
CompatHelper: bump compat for DynamicPolynomials to 0.5 for package test, (keep existing compat)
