+++ test/Project.toml
@@ -11,7 +11,7 @@ Test = "8dfed614-e22c-5e08-85e1-65c5234f0b40"
 [compat]
 AffineArithmetic = "0.2"
 Documenter = "0.27"
-DynamicPolynomials = "0.3, 0.4"
+DynamicPolynomials = "0.3, 0.4, 0.5"
 IntervalOptimisation = "0.4.1"
 SDPA = "0.2, 0.3, 0.4, 0.5"
 SumOfSquares = "0.3.6, 0.4, 0.5, 0.6"
