### 2023-07-01 2ac6234
bump SumOfSquares; use hyphen syntax
