### 2023-07-01 ecca6ed
update package versions; use hyphen syntax
