+++ Project.toml
@@ -11,7 +11,7 @@ Requires = "ae029012-a4dd-5104-9daa-d747884805df"
 
 [compat]
 ForwardDiff = "0.10"
-IntervalArithmetic = "0.15, 0.16, 0.17, 0.18, 0.19, 0.20"
+IntervalArithmetic = "0.15 - 0.20"
 ReachabilityBase = "0.1.1"
 Reexport = "0.2, 1"
 Requires = "0.5, 1"
