### 2023-08-12 db11ead
CompatHelper: bump compat for ReachabilityBase to 0.2, (keep existing compat)
