+++ Project.toml
@@ -12,7 +12,7 @@ Requires = "ae029012-a4dd-5104-9daa-d747884805df"
 [compat]
 ForwardDiff = "0.10"
 IntervalArithmetic = "0.15 - 0.20"
-ReachabilityBase = "0.1.1"
+ReachabilityBase = "0.1.1, 0.2"
 Reexport = "0.2, 1"
 Requires = "0.5, 1"
 julia = "1.6"
