+++ Project.toml
@@ -1,18 +1,16 @@
 name = "RangeEnclosures"
 uuid = "1b4d18b6-9e5d-11e9-236c-f792b01831f8"
-version = "0.1.4"
+version = "0.2.0"
 
 [deps]
 ForwardDiff = "f6369f11-7733-5829-9624-2563aa707210"
 IntervalArithmetic = "d1acc4aa-44c8-5952-acd4-ba5d80a2a253"
 ReachabilityBase = "379f33d0-9447-4353-bd03-d664070e549f"
-Reexport = "189a3867-3050-52da-a836-e630ba90ab69"
 Requires = "ae029012-a4dd-5104-9daa-d747884805df"
 
 [compat]
 ForwardDiff = "0.10"
 IntervalArithmetic = "0.15 - 0.21"
 ReachabilityBase = "0.1.1 - 0.2"
-Reexport = "0.2, 1"
 Requires = "0.5, 1"
 julia = "1.6"
