+++ test/univariate.jl
@@ -50,10 +50,20 @@ end
     rleft, rright = relative_precision(x, xref)
     @test rleft ≤ 1e-5 && rright ≤ 1e-5
 
-    x = enclose(p, dom, SumOfSquaresEnclosure(; backend=SDPA.Optimizer))
-    xref = Interval(4.8333, 10.541)
-    rleft, rright = relative_precision(x, xref)
-    @test rleft ≤ 1e-5 && rright ≤ 1e-5
+    if Sys.iswindows()
+        # SDPA is broken on Windows
+        @test_broken begin
+            x = enclose(p, dom, SumOfSquaresEnclosure(; backend=SDPA.Optimizer))
+            xref = Interval(4.8333, 10.541)
+            rleft, rright = relative_precision(x, xref)
+            rleft ≤ 1e-5 && rright ≤ 1e-5
+        end
+    else
+        x = enclose(p, dom, SumOfSquaresEnclosure(; backend=SDPA.Optimizer))
+        xref = Interval(4.8333, 10.541)
+        rleft, rright = relative_precision(x, xref)
+        @test rleft ≤ 1e-5 && rright ≤ 1e-5
+    end
 end
 
 @testset "Taylor-model solver without normalization" begin
