+++ test/runtests.jl
@@ -1,5 +1,11 @@
-using IntervalOptimisation, TaylorModels, Test, RangeEnclosures, SumOfSquares, SDPA,
-      AffineArithmetic
+using Test, RangeEnclosures
+using AffineArithmetic, IntervalOptimisation, TaylorModels, SumOfSquares
+
+@static if Sys.iswindows()
+    @test_broken using SDPA
+else
+    using SDPA
+end
 
 using DynamicPolynomials: @polyvar
 
@@ -13,8 +19,10 @@ available_solvers = (NaturalEnclosure(),
 include("univariate.jl")
 include("multivariate.jl")
 
-using Documenter
-include("../docs/init.jl")
-@time @testset "doctests" begin
-    doctest(RangeEnclosures)
+@static if !Sys.iswindows()  # broken due to SDPA
+    using Documenter
+    include("../docs/init.jl")
+    @time @testset "doctests" begin
+        doctest(RangeEnclosures)
+    end
 end
