### 2023-08-25 8592b59
CompatHelper: bump compat for IntervalArithmetic to 0.21, (keep existing compat)
