### 2023-09-11 #122 @schillic
Thanks all! This is very unfortunate timing :sweat_smile:

The `IntervalArithmetic` devs decided to change their API and I expect another major breaking change from them in the near future (the changes are already in their main branch). I would prefer to make the paper compatible with that new version before publishing. Concretely, they will remove `IntervalBox` and instead use `AbstractVector{<:Interval}`, which we might then do in this package too. I need to discuss this with @mforets and @lucaferranti.

@blegat @dpsanders: You can still take a look at the paper. The only thing I expect to change is the API for multivariate functions. So whenever you see `IntervalBox`, replace it with `AbstractVector{<:Interval}`.
