+++ src/intervalarithmetic.jl
@@ -2,7 +2,7 @@
 # Methods using interval arithmetic
 # =================================
 
-function _enclose(::NaturalEnclosure, f::Function, dom::Interval_or_IntervalBox; kwargs...)
+function _enclose(::NaturalEnclosure, f::Function, dom::Interval_or_IntervalVector; kwargs...)
     return f(dom)
 end
 
@@ -11,7 +11,7 @@ function _enclose(::MeanValueEnclosure, f::Function, dom::Interval;
     return f(mid(dom)) + df(dom) * (dom - mid(dom))
 end
 
-function _enclose(::MeanValueEnclosure, f::Function, dom::IntervalBox;
-                  df=t -> ForwardDiff.gradient(f, t.v))
+function _enclose(::MeanValueEnclosure, f::Function, dom::AbstractVector{<:Interval};
+                  df=t -> ForwardDiff.gradient(f, t))
     return f(mid.(dom)) + dot(df(dom), dom - mid.(dom))
 end
