+++ Project.toml
@@ -5,12 +5,14 @@ version = "0.2.1"
 [deps]
 ForwardDiff = "f6369f11-7733-5829-9624-2563aa707210"
 IntervalArithmetic = "d1acc4aa-44c8-5952-acd4-ba5d80a2a253"
+PkgVersion = "eebad327-c553-4316-9ea0-9fa01ccd7688"
 ReachabilityBase = "379f33d0-9447-4353-bd03-d664070e549f"
 Requires = "ae029012-a4dd-5104-9daa-d747884805df"
 
 [compat]
 ForwardDiff = "0.10"
 IntervalArithmetic = "0.15 - 0.21"
+PkgVersion = "0.3"
 ReachabilityBase = "0.1.1 - 0.2"
 Requires = "0.5, 1"
 julia = "1.6"
