+++ src/intervaloptimisation.jl
@@ -9,7 +9,7 @@ function _enclose(mse::MooreSkelboeEnclosure, f::Function, dom::Interval_or_Inte
     global_min, _ = minimise(f, dom; structure=mse.structure, tol=mse.tol)
     global_max, _ = maximise(f, dom; structure=mse.structure, tol=mse.tol)
 
-    return Interval(inf(global_min), sup(global_max))
+    return interval(inf(global_min), sup(global_max))
 end
 
 function load_intervaloptimization()
