### 2023-09-11 8d29490
add plot recipe for vector of Intervals
