+++ Project.toml
@@ -6,11 +6,13 @@ version = "0.2.1"
 ForwardDiff = "f6369f11-7733-5829-9624-2563aa707210"
 IntervalArithmetic = "d1acc4aa-44c8-5952-acd4-ba5d80a2a253"
 ReachabilityBase = "379f33d0-9447-4353-bd03-d664070e549f"
+RecipesBase = "3cdcf5f2-1ef4-517c-9805-6587b60abb01"
 Requires = "ae029012-a4dd-5104-9daa-d747884805df"
 
 [compat]
 ForwardDiff = "0.10"
 IntervalArithmetic = "0.15 - 0.21"
 ReachabilityBase = "0.1.1 - 0.2"
+RecipesBase = "0.6 - 0.8, 1"
 Requires = "0.5, 1"
 julia = "1.6"
