### 2023-09-11 #122 @schillic
About the `IntervalBox` ...  it is complicated ...
But, as said above, I think you can go ahead and review. Below is some background.

We have some optional dependencies, which we load in the tests. Four of these packages would need to be updated to the new `IntervalArithmetic` API before we can test the new version. I have not seen any effort in these packages for the past weeks, and most of these packages are not maintained for more than a year. So it is unlikely that we will get there, and so the user is probably stuck with the current version if these dependencies are used.

Whether we will support newer versions of `IntervalArithmetic` is currently unclear. I do not like to upgrade `IntervalArithmetic` and then only run the tests with the old `IntervalArithmetic` version - who knows whether things still work. But I also do not like to not test functionality with the optional dependencies. Most of the code in this package is actually about that. So there is a conflict. One way out could be to test basic functionality without the optional dependencies and with the latest `IntervalArithmetic` version, and then switching to another environment for the optional dependencies (or the other way around). But I am not aware of a good way to do that. I guess we would need to have different test folders/environments and different test scripts. But overengineering is rarely a good thing :)
