### 2023-09-11 #122 @dpsanders
I think we should move the current `IntervalBox` code out into a separate package. If you are able to do this then please feel free.
