### 2023-09-12 #122 @lucaferranti
> I think we should move the current IntervalBox code out into a separate package.

@dpsanders maybe open an issue on `IntervalArithmetic` about this?
