### 2023-09-13 #122 @blegat
The package looks really useful and the paper is clear and well motivated. I only have a few minor comments:

- You cite [5, 7] saying they provide algorithms but not specifying which and then you mention the branch and bound algorithm without reference. Is the branch and bound algorithm developed in [5, 7] ?
- "dependency problem and the wrapping effect": do you have any reference for this ?
- "where the enclosure obtained with plain interval arithmetic (natural enclosure)": use `\empth{natural enclosure}` as it's the first time we encounter this term so you are defining it.
- For [9], you can use the JuliaCon 2019 ref which is more recent
- Isn't it weird that the user needs to implement a method that starts with an underscore: `_enclose` ?
- "Intervaloptimisation.jl" -> the `o` should be capitalized
