+++ paper/ref.bib
@@ -1,8 +1,9 @@
-@inproceedings{legat2017sos,
-  author    = {Legat, Beno{\^\i}t and Coey, Chris and Deits, Robin and Huchette, Joey and Perry, Amelia},
-  title     = {{Sum-of-squares optimization in Julia}},
-  booktitle = {The First Annual JuMP-dev Workshop},
-  year      = {2017},
+@Conference{weisser2019polynomial,
+  author    = {Weisser, Tillmann and Legat, Beno{\^\i}t and Coey, Chris and Kapelevich, Lea and Vielma, Juan Pablo},
+  title     = {Polynomial and Moment Optimization in Julia and JuMP},
+  booktitle = {JuliaCon},
+  year      = {2019},
+  url       = {https://pretalx.com/juliacon2019/talk/QZBKAU/},
 }
 
 @inproceedings{benet2019taylormodels,
@@ -111,4 +112,11 @@
   publisher = {The Open Journal},
   url       = {https://doi.org/10.21105/jcon.00097},
   doi       = {10.21105/jcon.00097}
+}
+
+@book{neumaier1993wrapping,
+  title={The wrapping effect, ellipsoid arithmetic, stability and confidence regions},
+  author={Neumaier, Arnold},
+  year={1993},
+  publisher={Springer}
 }
