### 2023-09-29 cab0a03
Update paper/paper.tex
Co-authored-by: Christian Schilling <git@christianschilling.net>
