+++ paper/paper.tex
@@ -23,7 +23,7 @@ Motivated by this, we introduce \pkg{\RE}, a unified framework to bound the rang
 
 Given a function $f: \mathcal{D} \to \R$ over a domain $\mathcal{D}\subseteq\R$, the \emph{range} (or image) is the set $\mathcal{R}=\{y\in\R \mid \exists x\in\mathcal{D}: f(x)=y\}$. In practical applications, we are interested in determining the \textit{interval range} of $f$, i.e., the smallest interval containing $\mathcal{R}$. Unfortunately, computing the interval range of a multivariate function is NP-hard~\cite{kreinovich2002range}.
 
-For this reason, we practically seek an \textit{enclosure} $\mathcal{E}\supseteq\mathcal{R}$ of the interval range. A standard method to obtain an enclosure is to evaluate the function with interval arithmetic~\cite{moore2009introduction}, which however often produces a wide overestimation due to issues such as the dependency problem \cite{de2004affine} and the wrapping effect \cite{neumaier1993wrapping}. For this 
+For this reason, we practically seek an \textit{enclosure} $\mathcal{E}\supseteq\mathcal{R}$ of the interval range. A standard method to obtain an enclosure is to evaluate the function with interval arithmetic~\cite{moore2009introduction}, which however often produces a wide overestimation due to issues such as the dependency problem~\cite{de2004affine} and the wrapping effect~\cite{neumaier1993wrapping}. For this 
 reason, different algorithms have been developed over the past decades~\cite{de2004affine,hansen2003global}, but each comes with its own strengths and weaknesses. This is visualized in Fig.~\ref{fig:enclosure_2D}, where the enclosure obtained with plain interval arithmetic, called \emph{natural enclosure}, is compared to the result of a branch-and-bound algorithm. In general, when choosing an algorithm to compute a function enclosure, a trade-off between accuracy and computational efficiency has to be made.
 
 \smallskip
