+++ paper/paper.tex
@@ -50,11 +50,11 @@ The basic usage is via \texttt{enclose(f, D, solver; kwargs...)}, where \texttt{
 \end{figure}
 
 In \emph{\RE}, the \texttt{solver} is an instance of a struct that must be a subtype of \texttt{AbstractEnclosureAlgorithm}.
-If a user wants to add a new solver, they just have to add a new struct, say, \texttt{MyEnclosure} and implement the method \texttt{\_enclose}, which is automatically called by \texttt{enclose}.
+If a user wants to add a new solver, they just have to add a new struct, say, \texttt{MyEnclosure} and extend the method \texttt{enclose}.
 %
 \begin{lstlisting}[language=Julia]
-_enclose(solver::MyEnclosure, f::Function,
-         D::Union{Interval, IntervalBox}; kwargs...)
+enclose(f::Function, D::Union{Interval, IntervalBox},
+        solver::MyEnclosure; kwargs...)
 \end{lstlisting}
 
 Note that \texttt{D} can be of type \texttt{Interval} for univariate ($n = 1$) functions or of type \texttt{IntervalBox} for multivariate ($n > 1$) functions.
