### 2023-09-29 5fdf561
Update paper/ref.bib
Co-authored-by: Christian Schilling <git@christianschilling.net>
