### 2023-09-29 e222591
Update paper/paper.tex
Co-authored-by: Luca Ferranti <49938764+lucaferranti@users.noreply.github.com>
