+++ paper/paper.tex
@@ -50,7 +50,7 @@ The basic usage is via \texttt{enclose(f, D, solver; kwargs...)}, where \texttt{
 \end{figure}
 
 In \emph{\RE}, the \texttt{solver} is an instance of a struct that must be a subtype of \texttt{AbstractEnclosureAlgorithm}.
-If a user wants to add a new solver, they just have to add a new struct, say, \texttt{MyEnclosure} and extend the method \texttt{enclose}.
+If a user wants to add a new solver, they just have to add a new struct, say, \texttt{MyEnclosure} and extend the method \texttt{enclose} with the algorithm implementation, as the following code snippet demonstrates.
 %
 \begin{lstlisting}[language=Julia]
 enclose(f::Function, D::Union{Interval, IntervalBox},
