+++ paper/paper.tex
@@ -53,8 +53,10 @@ In \emph{\RE}, the \texttt{solver} is an instance of a struct that must be a sub
 If a user wants to add a new solver, they just have to add a new struct, say, \texttt{MyEnclosure} and extend the method \texttt{enclose} as the following code snippet demonstrates.
 %
 \begin{lstlisting}[language=Julia]
-enclose(f::Function, D::Union{Interval, IntervalBox},
-        solver::MyEnclosure; kwargs...)
+function enclose(f::Function, dom::Union{Interval,IntervalBox},
+         solver::MyEnclosure; kwargs...)
+    # Write your algorithm implementation here.
+end
 \end{lstlisting}
 
 Note that \texttt{D} can be of type \texttt{Interval} for univariate ($n = 1$) functions or of type \texttt{IntervalBox} for multivariate ($n > 1$) functions.
