+++ paper/paper.tex
@@ -55,7 +55,7 @@ If a user wants to add a new solver, they just have to add a new struct, say, \t
 \begin{lstlisting}[language=Julia]
 function enclose(f::Function, dom::Union{Interval,IntervalBox},
          solver::MyEnclosure; kwargs...)
-    # Write your algorithm implementation here.
+    # Write your algorithm here.
 end
 \end{lstlisting}
 
