### 2023-09-30 3cc2aa1
Update paper/paper.tex
Co-authored-by: Christian Schilling <git@christianschilling.net>
