### 2023-09-30 3634936
Apply suggestions from code review
Co-authored-by: Christian Schilling <git@christianschilling.net>
