### 2023-09-30 a3b66c5
Update paper/paper.tex
Co-authored-by: Christian Schilling <git@christianschilling.net>
