+++ paper/ref.bib
@@ -120,3 +120,22 @@
   year={1993},
   publisher={Springer}
 }
+
+@article{karhbet2017range,
+  title={Range bounds of functions over simplices, for branch and bound algorithms},
+  author={Karhbet, Sam and Kearfott, Ralph Baker},
+  journal={Reliable Comput},
+  volume={25},
+  number={7},
+  pages={53--73},
+  year={2017}
+}
+
+@incollection{vzilinskas2007survey,
+  title={A survey of methods for the estimation ranges of functions using interval arithmetic},
+  author={{\v{Z}}ilinskas, Julius and Bogle, Ian David Lockhart},
+  booktitle={Models and Algorithms for Global Optimization: Essays Dedicated to Antanas {\v{Z}}ilinskas on the Occasion of His 60th Birthday},
+  pages={97--108},
+  year={2007},
+  publisher={Springer}
+}
