### 2023-09-30 9311a9d
fix multivariate branch/bound for IA v0.21
