+++ src/branchandbound.jl
@@ -44,11 +44,11 @@ function _monotonicity_check(f::Function, X::IntervalBox{N}, ∇fX::AbstractVect
 
     @inbounds for (i, di) in enumerate(∇fX)
         if di >= 0  #  increasing
-            high[i] = sup(X[i])
-            low[i] = inf(X[i])
+            high[i] = interval(sup(X[i]))
+            low[i] = interval(inf(X[i]))
         elseif di <= 0  # decreasing
-            high[i] = inf(X[i])
-            low[i] = sup(X[i])
+            high[i] = interval(inf(X[i]))
+            low[i] = interval(sup(X[i]))
         else
             return di, false
         end
