### 2023-09-30 6d92aeb
Apply suggestions from code review
Co-authored-by: Christian Schilling <git@christianschilling.net>
