+++ paper/paper.tex
@@ -34,7 +34,7 @@ We present \pkgl{https://github.com/JuliaReach/RangeEnclosures.jl}{\RE}, a Julia
 
 \section{A tour through \RE}
 
-In this section we give a quick overview of the API to bound function ranges. The package offers several solvers for this purpose, such as natural (interval) enclosure, mean-value form~\cite{moore2009introduction}, Moore-Skelboe algorithm~\cite{hansen2003global}, branch-and-bound \cite{karhbet2017range} and Taylor models~\cite{benet2019taylormodels}, or polynomial optimization~\cite{weisser2019polynomial}. The full list of implemented solvers can be found in the package documentation\footnote{\url{https://juliareach.github.io/RangeEnclosures.jl/}}.
+In this section we give a quick overview of the API to bound function ranges. The package offers several solvers for this purpose, such as natural (interval) enclosure, mean-value form~\cite{moore2009introduction}, Moore-Skelboe algorithm~\cite{hansen2003global}, branch-and-bound~\cite{karhbet2017range} and Taylor models~\cite{benet2019taylormodels}, or polynomial optimization~\cite{weisser2019polynomial}. The full list of implemented solvers can be found in the package documentation\footnote{\url{https://juliareach.github.io/RangeEnclosures.jl/}}.
 
 
 \subsection{The enclose API}
