+++ paper/ref.bib
@@ -124,7 +124,7 @@
 @article{karhbet2017range,
   title={Range bounds of functions over simplices, for branch and bound algorithms},
   author={Karhbet, Sam and Kearfott, Ralph Baker},
-  journal={Reliable Comput},
+  journal={Reliab. Comput.},
   volume={25},
   number={7},
   pages={53--73},
@@ -137,5 +137,6 @@
   booktitle={Models and Algorithms for Global Optimization: Essays Dedicated to Antanas {\v{Z}}ilinskas on the Occasion of His 60th Birthday},
   pages={97--108},
   year={2007},
-  publisher={Springer}
+  publisher={Springer},
+  doi={10.1007/978-0-387-36721-7\_6}
 }
