+++ paper/ref.bib
@@ -118,7 +118,8 @@
   title={The wrapping effect, ellipsoid arithmetic, stability and confidence regions},
   author={Neumaier, Arnold},
   year={1993},
-  publisher={Springer}
+  publisher={Springer},
+  doi = {10.1007/978-3-7091-6918-6\_14}
 }
 
 @article{karhbet2017range,
