### 2023-09-30 #122 @lucaferranti
@blegat we have now addressed your review comments. Here is a short summary

> You cite [5, 7] saying they provide algorithms but not specifying which and then you mention the branch and bound algorithm without reference. Is the branch and bound algorithm developed in [5, 7] ?

The purpose of the sentence in the introduction was simply to offer a generic motivation and give a more detailed algorithm - reference corresponse in section 2. To avoid confusion, we now added a survey paper on range enclosures methods and cite that in the introduction In section 2, we have a reference for every mentioned algorithm. We have also added a reference for the branch and bound.

> "dependency problem and the wrapping effect": do you have any reference for this ?

reference has been added

> "where the enclosure obtained with plain interval arithmetic (natural enclosure)": use \empth{natural enclosure} as it's the first time we encounter this term so you are defining it.

done

> For [9], you can use the JuliaCon 2019 ref which is more recent

updated

> Isn't it weird that the user needs to implement a method that starts with an underscore: _enclose ?

You are right that the `_enclose` internal method was an extra layer of complexity. The API has now been restructured so that users who want to add new algorithms can simply extend `RangeEnclosures.enclose`. The paper has beeen updated to reflect this.

> "Intervaloptimisation.jl" -> the o should be capitalized

fixed.

Please take a look at the new paper and let us know if you are happy with the new version.
