+++ paper/paper.tex
@@ -62,7 +62,7 @@ function enclose(f::Function,
     # solver-specific implementation
 end
 \end{lstlisting}
-\vspace*{-2mm}
+\vspace*{-3mm}
 
 Note that \texttt{D} can be of type \texttt{Interval} for univariate ($n = 1$) functions or of type \texttt{IntervalBox} for multivariate ($n > 1$) functions.
 
