### 2023-09-30 #122 @schillic
> I think we should move the current `IntervalBox` code out into a separate package. If you are able to do this then please feel free.

@dpsanders: We do not plan to work around the planned `IntervalBox` removal and rather freeze the `IntervalArithmetic` version as it is. This is simply because almost all backends have not even started supporting `IntervalArithmetic` v0.21, so we see little value in trying to support the next breaking version if it would not be usable.

Hence from our side the paper is stable.
