### 2023-10-07 0f45c41
CompatHelper: bump compat for IntervalArithmetic to 0.21 for package docs, (keep existing compat)
