+++ docs/Project.toml
@@ -6,6 +6,6 @@ Plots = "91a5bcdd-55d7-5caf-9e0b-520d859cae80"
 
 [compat]
 Documenter = "1"
-IntervalArithmetic = "0.20"
+IntervalArithmetic = "0.20, 0.21"
 IntervalOptimisation = "0.4"
 Plots = "1"
