### 2023-10-07 47f468d
freeze IntervalArithmetic v0.20 in /docs
