### 2023-10-08 400576e
CompatHelper: bump compat for IntervalArithmetic to 0.21 for package docs, (keep existing compat)
