### 2023-10-15 15c013f
use latest supported version of IntervalArithmetic in docs
