### 2023-10-28 #122 @dpsanders
Apologies for the delay in reviewing this.

This is a nice package that brings together in a neat way various different range enclosure methods from a variety of packages in the Julia ecosystem.

Even though this package does not contain that much code, a user could have a difficult time learning to use each of the packages that it wraps, so this is very useful.

The paper is well-written and well-motivated, and the docs are concise and useful.

A couple of comments:
- Maybe you could add an explicit example of how to extend the `enclose` function with a new method.
- It could be helpful to indicate the computational complexity of each method (although I appreciate that this is non-trivial!)
- You seem to have implemented your own branch-and-bound mechanism in the package. How does this differ from e.g. IntervalOptimization.jl?
- The [EAGO package](https://github.com/PSORLab/EAGO.jl) is a much more sophisticated interval optimization package, which could be useful to wrap as well.

Otherwise I am happy to see this paper published in JOSS!
