### 2023-10-29 #122 @lucaferranti
Thank you for the review @dpsanders , please do make sure to also tick the boxes above :)

> You seem to have implemented your own branch-and-bound mechanism in the package. How does this differ from e.g. IntervalOptimization.jl?

The main difference is that in the Moore-Skelboe algorithm one **throws away** boxes which are proved to not contain an extremum. The implemented branch-and-bound, instead, aggregates boxes into a range extimator as you bisect.

Roughly,

- Start with a range estimation`Y` (initialized to empty interval)
- Consider the subdomain `Xi`, if you can prove the function is monotone over `Xi`, then no need to bisect, evaluate the function at the end points and update the range estimate to `Y \cup f(Xi)`.
- If you cannot prove monotonocity ($$f'(Xi)$$ contains a $$0$$), then bisect the box and go on (until the tolerance for stopping bisecting is met).

The strategy is very similar to a triditional branch-and-bound, I guess this is technically more a branch-and-accumulate

> Maybe you could add an explicit example of how to extend the enclose function with a new method.

This is done in the code snippet at the beginning of section 2.1, any thing in particular you find unclear there that we should clarify?

> The [EAGO package](https://github.com/PSORLab/EAGO.jl) is a much more sophisticated interval optimization package, which could be useful to wrap as well.

I'll look into it, thanks! I've opened an issue to discuss this

> It could be helpful to indicate the computational complexity of each method (although I appreciate that this is non-trivial!)

Indeed, it can be a bit challenging, I am also not sure with respect to what measure the complexity here. With respect to the input dimension? I would expect roughly all of those to suffer from the curse of dimensionality which is typical for inteval methods, but I'll see if we can add some insights about complexity of the algorithms in the paper or documentation.
