### 2023-10-29 JuliaReach/RangeEnclosures.jl#149 @dpsanders
Wrong answer for 2D example in docs
```
julia> h(x) = sin(x[1]) - cos(x[2]) - sin(x[1]) * cos(x[1])
h (generic function with 1 method)

julia> Dh = IntervalBox(-5..5, -5..5)
[-5.0, 5.0]²

julia> Rh = enclose(h, Dh, BranchAndBoundEnclosure())
[-0.970576, 0.403252]
```
This is different from what's in the docs and is completely wrong.

```

(@v1.9) pkg> st
Status `~/.julia/environments/v1.9/Project.toml`
⌃ [79e6a3ab] Adapt v3.6.2
  [6e4b80f9] BenchmarkTools v1.3.2
  [5789e2e9] FileIO v1.16.1
  [fb4d412d] FixedPointDecimals v0.4.3
  [e9467ef8] GLMakie v0.8.11
  [2cbf6e9d] HardSphereDynamics v0.1.3 `~/Dropbox/packages/HardSphereDynamics`
  [d1acc4aa] IntervalArithmetic v0.21.2
  [ee78f7c6] Makie v0.19.11
  [5fb14364] OhMyREPL v0.5.23
  [91a5bcdd] Plots v1.39.0
⌃ [c3e4b0f8] Pluto v0.19.29
  [9c30249a] RAI v0.2.8
  [1b4d18b6] RangeEnclosures v0.2.2
⌃ [295af30f] Revise v3.5.6
  [90137ffa] StaticArrays v1.6.5
⌅ [2913bbd2] StatsBase v0.33.21
```

Version info:
```
  OS: macOS (arm64-apple-darwin22.4.0)
  CPU: 12 × Apple M2 Max
  WORD_SIZE: 64
  LIBM: libopenlibm
  LLVM: libLLVM-14.0.6 (ORCJIT, apple-m1)
```
