### 2023-10-29 JuliaReach/RangeEnclosures.jl#149 @lucaferranti
interesting, this seems to be due to some changes introduced in IA v0.21

```julia
(jl_Nk2nko) pkg> st
Status `/tmp/jl_Nk2nko/Project.toml`
  [d1acc4aa] IntervalArithmetic v0.21.2
  [1b4d18b6] RangeEnclosures v0.2.2

julia> using RangeEnclosures

julia> using IntervalArithmetic

julia> h(x) = sin(x[1]) - cos(x[2]) - sin(x[1]) * cos(x[1])
h (generic function with 1 method)

julia> Dh = IntervalBox(-5 .. 5, -5 .. 5)

[-5.0, 5.0]²

julia> res = enclose(h, Dh, BranchAndBoundEnclosure())
[-0.970576, 0.403252]
```

Because using 0.20 it works as expected

```julia
(jl_NwKnpk) pkg> st
Status `/tmp/jl_NwKnpk/Project.toml`
⌃ [d1acc4aa] IntervalArithmetic v0.20.9
  [1b4d18b6] RangeEnclosures v0.2.2
Info Packages marked with ⌃ have new versions available and may be upgradable.

julia> using RangeEnclosures, IntervalArithmetic
[ Info: Precompiling RangeEnclosures [1b4d18b6-9e5d-11e9-236c-f792b01831f8]

julia> h(x) = sin(x[1]) - cos(x[2]) - sin(x[1]) * cos(x[1])
h (generic function with 1 method)

julia> Dh = IntervalBox(-5 .. 5, -5 .. 5)
[-5, 5]²

julia> res = enclose(h, Dh, BranchAndBoundEnclosure())
[-2.71068, 2.71313]
```

the reason why this goes undetected on the unit tests is because on CI we use IA 0.20 (because some optional dependencies, e.g. TaylorModels.jl I believe, do not support yet IA 0.21
