### 2023-10-29 2e144dd
fix comparison to work with 0.21
