### 2023-10-30 #122 @dpsanders
Thanks @lucaferranti, I missed the snippet in section 2.1.
If that's not in the actual package docs, then please add it there too.

Thanks for fixing the [correctness issue](https://github.com/JuliaReach/RangeEnclosures.jl/issues/149) I reported in the repo. Has a new version of the package been released that contains that fix?
