### 2023-11-24 b77f107
fix variable name in paper
