### 2023-11-24 #122 @schillic
@dpsanders:
- Yes, the fix has been published.
- Yes, we added the snippet to the documentation as well ([here](https://juliareach.github.io/RangeEnclosures.jl/dev/tutorial/#Adding-a-new-enclosure-algorithm)).
- It seems that these complexity results are not so easy to grasp. Do you think it is crucial to include them to publish the paper? I understood this more as a suggestion.
