### 2023-12-14 0735f90
CompatHelper: bump compat for IntervalArithmetic to 0.22, (keep existing compat)
