### 2023-12-19 4bcf64e
freeze version of IntervalArithmetic
