### 2024-01-11 #122 @odow
@schillic As a final step, can you make a Zenodo (https://zenodo.org) release and then post the DOI here? You probably need to make a new GitHub release for it as well.
