### 2024-01-13 #122 @whedon
:wave: @JuliaCon/jcon-eics, this paper is ready to be accepted and published.
