### 2024-03-15 #122 @lucaferranti
following messages to test new editorial bot infrastructure
